backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pymongo==4.6.1
motor==3.3.2
pydantic==2.5.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sentence-transformers==2.3.1
faiss-cpu==1.7.4
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
httpx==0.26.0
beautifulsoup4==4.12.3
feedparser==6.0.11
redis==5.0.1
